Migrate cli to .ts specifiers, externalize deps
The CLI still used .js relative specifiers, so running it straight from
source failed with ERR_MODULE_NOT_FOUND -- Node cannot resolve a .js
specifier pointing at a .ts file. It only worked because esbuild and
Vitest resolve that themselves. Switching to .ts gives the CLI the
treatment the other packages already got, so devs and agents can run it
with no build step.

The bundle also declared every runtime dependency and then inlined it,
so an installer downloaded a 744KB bundle plus the whole viem tree and
used neither copy. Declared deps are external now, dropping the bundle
to 15KB. The externals rule and the private-workspace guard move to
scripts/package-externals.js, shared with bundle-package.js. The CLI
keeps its own esbuild config: it bundles a bin rather than exports, and
needs platform node for the node: builtins in program.ts.

esbuild.ts becomes esbuild.js because tsc typechecks it and it now
imports an untyped JS helper. The createRequire banner is dropped --
commander 15 is pure ESM and nothing third-party is inlined anymore.
Bundling hangs off postbuild so CI exercises it, matching #712.
