+++ packages/cli/test/e2e/helpers.ts
@@ -35,9 +35,9 @@ import {
 } from "viem/actions";
 import { mainnet } from "viem/chains";
 
-import { type GlobalOptions } from "../../src/lib/client.js";
-import { printError } from "../../src/lib/output.js";
-import { createProgram } from "../../src/program.js";
+import { type GlobalOptions } from "../../src/lib/client.ts";
+import { printError } from "../../src/lib/output.ts";
+import { createProgram } from "../../src/program.ts";
 
 export const usdc = {
   address: "0xA0b86991c6218b36c1d19D4a2e9Eb0cE3606eB48",
